Correct the ui
