+++ README.md
@@ -85,9 +85,9 @@ Before to run the program, verify that you have Ruby installed on your OS runnin
 
 Run the following command inside the project folder:
 1. Open your terminal and navigate to the root directory of your project.
-2. Run `IRB`
-3. require the files you want to use. Ex: `require "./app.rb"`
-4. Now you are ready
+2. Open the `lib` directory ```$ cd lib```
+3. Run ```$ ruby main.rb```
+4. Enjoy
 
 <p align="right">(<a href="#readme-top">back to top</a>)</p>
