+++ lib/app/book.rb
