+++ lib/app/capitalize_decorator.rb
