+++ lib/app/classroom.rb
