+++ lib/app/edit_name.rb
