+++ lib/app/modules/rental_mod.rb
