+++ lib/app/nameable.rb
