+++ lib/app/rental.rb
