+++ lib/app/student.rb
