+++ lib/app/teacher.rb
