+++ lib/app/trimmer_decorator.rb
