--- school_lib.rb
@@ -1,8 +1,0 @@
-require './lib/app'
-
-def main
-  app = App.new
-  app.main_menu
-end
-
-main
